feat: add theme provider and split button variants
