+++ frontend/src/components/theme-provider.tsx
@@ -1,0 +1,9 @@
+import * as React from "react"
+
+type ThemeProviderProps = {
+  children: React.ReactNode
+}
+
+export function ThemeProvider({ children }: ThemeProviderProps) {
+  return <>{children}</>
+}
